Move BackendBenchmark, BdvHandleService to demo package
